Some functional content
Was working, but I didn't check this commit so it's probably broken. Slowly shifting implementation-specific stuff to the actual implementing classes instead of the abstract classes, which is tricky since I still want to make abstract classes that are as close to useful as possible. The hardest part is going to be a rewrite for Model that accepts things that don't just look like OpenGL. It can look like OpenGL in Dumbbell, but by the time it hits the abstract Model class it should be more generalized.