Added (and use) Avalanche quick starter function
The new start() function is used to make main() as quick and dirty as possible. Essentially, all main should do is:
1. Handle program arguments (if necessary)
2. Call Engine.start() with appropriate arguments
Unfortunately, static methods get a bit awkward across abstract classes and their subclasses. The solution here was to have methods with identical signatures in the super and sub classes and then have the abstract version throw a NotImplementedException. That way, if anything calls engine.start(...) and the implemented engine does *not* contain an overridden start method, the program will error out.