Updated Avalanche/Base components
Added typing to MVC components (and engine) to allow for more fluid communication between parts that know their companions. Also made it so that strictly Avalanche components and extensions could work on Avalanche (and with each other). Also built out the game logic for AvModel and created vertex/shader programs and some extra utilities to house vertices (and, thus, polygons).